+++ model/User.js
@@ -31,8 +31,16 @@ module.exports = class User extends Base {
                 ['email', 'email'],
                 [['blocked', 'verified', 'expiredPassword'], 'checkbox'],
                 ['unlockAt', 'date'],
-                ['name', 'unique', {ignoreCase: true, skipOnAnyError: true, message: 'auth.nameAlreadyTaken'}],
-                ['email', 'unique', {ignoreCase: true, skipOnAnyError: true, message: 'auth.emailAlreadyTaken'}]
+                ['name', 'unique', {
+                    ignoreCase: true,
+                    skipOnAnyError: true,
+                    message: 'auth.nameAlreadyTaken'
+                }],
+                ['email', 'unique', {
+                    ignoreCase: true,
+                    skipOnAnyError: true,
+                    message: 'auth.emailAlreadyTaken'
+                }]
             ],
             BEHAVIORS: {
                 'timestamp': require('areto/behavior/TimestampBehavior')